requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
distro==1.9.0
Django==5.2.1
django-environ==0.12.0
exceptiongroup==1.3.0
ffmpeg-python==0.2.0
filelock==3.18.0
fsspec==2025.3.2
future==1.0.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
llvmlite==0.44.0
Markdown==3.8
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.5
openai==1.78.1
openai-whisper==20240930
packaging==25.0
pydantic==2.11.4
pydantic_core==2.33.2
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.14.0
tiktoken==0.9.0
torch==2.7.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
yt-dlp==2025.4.30
